Revert "moved the files in tests&updated gitignore"
