+++ src/tests/service/DepositTests.java
@@ -1,4 +1,4 @@
-package tests.service.transaction;
+package tests.service;
 
 import model.*;
 import repository.*;
